+++ crates/plato/src/main.rs
@@ -3,6 +3,7 @@
 
 mod app;
 mod constants;
+mod helpers;
 mod task;
 
 use crate::app::run;